requirements.txt
canonicalwebteam.flask-base==0.3.1
canonicalwebteam.http==1.0.1
Flask==1.1.1
python-frontmatter==0.4.5
mistune==0.8.4
bleach==3.1.0
talisker[gunicorn]==0.15.0

